apps/dashboard/requirements.txt
flask>=3.1
paho-mqtt>=2.1.0
